Remove useless signing test
This test does too many things wrong:

It does `ok(encode_base64($xml), $expect)` which will return a true value but
even if we change it is essentially a string by string comparision of XML.

If we are going to use `is()` a change in the XML, eg
`<SignatureValue>\n$value\n</SignatureValue>` will render the test useless.

Coverage before and after stay the same, so I'm not going to spent time on
getting this test to work in a correct way.

Signed-off-by: Wesley Schwengle <waterkip@cpan.org>